requirements.txt
openai
langchain
chromadb
gitpython
tiktoken